Refactor VM controller folders
This patch moves the controllers related to the VM API (ex. the
volume and VM controllers) under controllers/virtualmachine.
