+++ controllers/virtualmachine/controllers.go
@@ -1,0 +1,26 @@
+// Copyright (c) 2024 VMware, Inc. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package virtualmachine
+
+import (
+	"fmt"
+
+	"sigs.k8s.io/controller-runtime/pkg/manager"
+
+	"github.com/vmware-tanzu/vm-operator/controllers/virtualmachine/virtualmachine"
+	"github.com/vmware-tanzu/vm-operator/controllers/virtualmachine/volume"
+	pkgctx "github.com/vmware-tanzu/vm-operator/pkg/context"
+)
+
+// AddToManager adds the controllers to the provided manager.
+func AddToManager(ctx *pkgctx.ControllerManagerContext, mgr manager.Manager) error {
+	if err := virtualmachine.AddToManager(ctx, mgr); err != nil {
+		return fmt.Errorf("failed to initialize virtualmachine controller: %w", err)
+	}
+	if err := volume.AddToManager(ctx, mgr); err != nil {
+		return fmt.Errorf("failed to initialize virtualmachine volume controller: %w", err)
+	}
+
+	return nil
+}
