+++ controllers/virtualmachine/volume/volume_controller_suite_test.go
@@ -10,7 +10,7 @@ import (
 
 	ctrlmgr "sigs.k8s.io/controller-runtime/pkg/manager"
 
-	"github.com/vmware-tanzu/vm-operator/controllers/volume"
+	"github.com/vmware-tanzu/vm-operator/controllers/virtualmachine/volume"
 	pkgcfg "github.com/vmware-tanzu/vm-operator/pkg/config"
 	pkgctx "github.com/vmware-tanzu/vm-operator/pkg/context"
 	providerfake "github.com/vmware-tanzu/vm-operator/pkg/providers/fake"
